Move channels into submodules
